Update .gitignore
opened the project up on my desktop. had to change where the project working directory was to allow the files to be read in properly with the path I gave the scanner.

IntelliJ:
project directory CSCI-446-Art-Int
project 1 folder inside that that contains the src folder.

Since csci-446 is the main directory then the following path will work:
project 1/sudoku puzzles/

cause a new .iml file for the new project space so added it to .gitignore