Local Search Simulated Annealing
LocalSearch - Started to layout simulated annealing (SA) function. Added stuff to main to test SA.

ConstraintSolver - I assume the classes initialize puzzles so I added that to the constraintSolver.

PuzzleImporter - Added a way to set the sudokuPuzzles to something new. Also made an empty function to getNeighbors but Im not sure if it is needed.