Validation, 3x3 Tracking, Neighbors, Local Search Simulated Annealing
Added a way to check for conflicts in puzzle.

Added a way to get the neighbors of a space.

Continued Implementation of Local Search Simulated Annealing. Working on deltaE and boltzmann probability next.

Added a way to see what 3x3 cube your space is assigned in.

Added a way to get all the neighbors of a space.