BayesianNetwork Set-Up
Implemented the BayesianNetwork from the .bif files

BayesianNetwork contains two dictionaries: variableDict, probabilityDict

variableDict contains all of the nodes of the graph in the form of a Variable class. Each variable contains a list of its parents keys.

probabilityDict contains all of the conditionalProbabilityTables (CPT) for a conditional probability. The key is the general conditional probability. Example keys from child.bif, "BirthAsphyxia" and "HypDistrib | DuctFlow, CardiacMixing". The returned object is a Probability class . These CPTs are stored in a dictionary within the Probability class where the key is the given evidence value (Example for "Birth Asphyxia" would be Key "table". for "HypeDistrib | . . ." would be key "Lt_to_Rt, None" and the returned object is a Double[] with the probabilities.

All data held in these classes are public for easy access.